Remove CdkDropList dependency because we already import DragDropModule. Rename file to add component in the name.
