+++ src/assets/wise5/components/match/match-student/match-student.module.ts
@@ -4,7 +4,7 @@ import { AddMatchChoiceDialogComponent } from './add-match-choice-dialog/add-mat
 import { MatchStudent } from './match-student.component';
 import { StudentComponentModule } from '../../../../../app/student/student.component.module';
 import { MatchCommonModule } from '../match-common.module';
-import { MatchStudentChoiceReuseComponent } from './match-student-choice-reuse/match-student-choice-reuse';
+import { MatchStudentChoiceReuseComponent } from './match-student-choice-reuse/match-student-choice-reuse.component';
 import { MatchStudentDefaultComponent } from './match-student-default/match-student-default.component';
 
 @NgModule({
